requirements.txt
Django==4.0.1
djangorestframework==3.13.1
django-seed==0.3.1
celery==5.0.5
redis==4.1.1